Don't pass MAC address to tapclient
Additionally if we are running in the context of Rancher we can't use
DNS because Rancher will not inject DNS to VMs.  This means we use the
rancher-metadata IP and not the DNS name.  As a consequence we also
don't need to wait for metadata to be available.  It will always be
available on start.