Remove vm-base docker image, in prefer to use VM image as lower layer of container
It's much easier to update only `startvm` script when we're adding more
feature/bug fixes to it. The VM image itself is relative stable. It's waste
of time to push the same VM image to Docker hub if it's not changed.

So now I remove the vm-base container, and use --build-arg parameter of
`docker build` to include VM image instead. This will put VM image at lower
layer of container, and it won't be upload if image itself doesn't changed.